cherry-pick 6aadb0a
[builtins] Fix TypedArray constructor case for array like argument

The ConstructByArrayLike case didn't handle JSArray with hole values
properly on the fast path. Even though the JSArray is a fast one and
Array prototype and Symbol.iterator in untouched the hole values must
be converted to undefined because conversion of a value to number
might modify the Array's prototype chain and thus alter the result of
reading a hole from a temporary JSArray object created as a part of
IteratorToList step.

Bug: chromium:1470560
Change-Id: I6395df541aa41675df63c9d8dad86b0201cf8fed
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4852525
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#89927}